// emasaku-cloudflare--semua-menjadi-satu/types.ts
export enum UserRole {
  MEMBER = 'Member',
  ADMIN = 'Admin',
  DEVELOPER = 'Developer',
}

export interface User {
  id: string;
  name: string;
  email?: string;
  avatarUrl: string;
  role: UserRole;
  joinedDate: string;
  whatsapp?: string;
  driverBalance?: number;
  isRetailPartner?: boolean;
  isLaundryPartner?: boolean;
  isSeller?: boolean;
  isWastePartner?: boolean;
}
// ...
export type RequestStatus = 'pending' | 'approved' | 'rejected';

export interface BaseRegistrationRequest {
    id: string;
    date: string;
    userId: string;
    userName: string;
    userAvatar: string;
    status: RequestStatus;
}

export interface SellerRegistrationRequest extends BaseRegistrationRequest {
    storeName: string;
    storeDescription: string;
    pickupAddress: string;
}

export interface DriverRegistrationRequest extends BaseRegistrationRequest {
    vehicleType: 'Motor' | 'Mobil';
    licensePlate: string;
    vehicleModel: string;
}

export interface ArisanRegistrationRequest extends BaseRegistrationRequest {
    groupName: string;
    feeAmount: number;
    paymentPeriod: 'Mingguan' | 'Bulanan';
    memberCount: number;
    startDate: string;
    description?: string;
}

export interface WastePartnerRegistrationRequest extends BaseRegistrationRequest {
    bankName: string;
    address: string;
    operationalHours: string;
    acceptedWastes: string;
}

export interface LaundryPartnerRegistrationRequest extends BaseRegistrationRequest {
    businessName: string;
    address: string;
    hours: string;
    services: {
        kiloan: boolean;
        satuan: boolean;
        ekspres: boolean;
        setrika: boolean;
    };
}

export interface RetailPartnerRegistrationRequest extends BaseRegistrationRequest {
    storeName: string;
    address: string;
}